年度計畫報告 : RA026 封面, RA027 目錄, RA028 工作計畫

+++ DomainStorm.Project.TWCrepair.Report.Web/Services/Impl/Staging/RA024Service.cs
@@ -99,7 +99,7 @@ public class RA024Service : IGetService<RA024, string>
         {
             PipeKind = x.PipeCode,
             PipeDiameter = x.Width.ToString(),
-            Length = x.Length
+            Length = (int)x.Length
         });
         result.GenerateData(fixForms, importPipes);
